% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gb_companies.R
\name{gb_companies}
\alias{gb_companies}
\title{GiantBomb Companies}
\usage{
gb_companies(id = NULL, n = 100, field_list = NULL, limit = 100,
  offset = 0, sort = NULL, filter = NULL, user_agent = NULL,
  delay = 30, api_key = gb_key())
}
\arguments{
\item{id}{an object id.}

\item{n}{total number of results returned. Defaults to 100.}

\item{field_list}{columns to return.}

\item{limit}{total results per page. Defaults to 100.}

\item{offset}{starting result.}

\item{sort}{arrange results by column. Accepts a string with the format
"column:direction".}

\item{filter}{reduce results by column attribute. Accepts a string with the
format "column:value".}

\item{user_agent}{allows you to customize the user agent for your requests.
See the Details section for more information.}

\item{delay}{the wait time between each query.}

\item{api_key}{a GiantBomb api key. The default searches for GB_KEY in your
environment variables.}
}
\description{
Retrieve a company or companies from the GiantBomb API.
}
\details{
An api key can be retrieved for the GiantBomb API after signing up
at \url{https://giantbomb.com/api}.

GiantBomb limits the query rate to 200 requests per resource per hour.
Additionally, you may experience issues if too many queries are made every
second. Each function attempts to mitigate problems by using a delay after
each query (default is 30 seconds). You can set \code{delay} to 0 if you want to
get the information faster, but GiantBomb may begin to block your requests
if you do this too often.

You may also want to use the \code{user_agent} parameter to customize how your
queries are being received. Doing so can minimize the possibility of
getting a captcha challenges.

Each query returns a maximum of 100 results per request. In order to get
1000 results with a \code{limit} of 100, it would take 10 api requests.
}
\examples{
\dontrun{
# Get 10 companies.
comp <- gb_companies(n = 10)

# Get a single company with an id.
comp <- gb_companies(id = 360)

# Use filters and sorting.
comp <- gb_companies(filter = "name:2k", sort = "date_added:asc")
}

}
