% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rbind2}
\alias{rbind2}
\title{rbind2}
\usage{
rbind2(...)
}
\arguments{
\item{...}{two or more valid objects that can be joined with \code{rbind}.}
}
\description{
Bind rows with mismatching column names.
}
\note{
\code{dplyr::bind_rows} is a much better alternative, but implementing
\code{rbind2} kept the number of package dependencies low.
}
