% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_offset}
\alias{set_offset}
\title{Set Offset}
\usage{
set_offset(offset, n, limit)
}
\arguments{
\item{offset}{starting position for the results.}

\item{n}{total number of results.}

\item{limit}{number of results per query.}
}
\description{
Given an n-size, limit, and offset parameter, return a vector of offsets.
}
